+++ src/main/scala/se/lu/nateko/cp/data/streams/geo/PointReducerState.scala
@@ -5,7 +5,7 @@ import scala.collection.mutable.Map
 
 class PointReducerState{
 
-	val lats, lons = Buffer.empty[Float]
+	val lats, lons = Buffer.empty[Double]
 	val bbox = new BBox
 	val shortList = Buffer.empty[Int]
 
@@ -18,25 +18,25 @@ class PointReducerState{
 class BBox(var left: Point, var top: Point, var right: Point, var bottom: Point){
 	def this() = this(new Point(1000, 0), new Point(0, -1000), new Point(-1000, 0), new Point(0, 1000))
 
-	def updateWith(lon: Float, lat: Float): Unit = {
+	def updateWith(lon: Double, lat: Double): Unit = {
 		if(lon < left.lon) left = new Point(lon, lat)
 		if(lon > right.lon) right = new Point(lon, lat)
 		if(lat < bottom.lat) bottom = new Point(lon, lat)
 		if(lat > top.lat) top = new Point(lon, lat)
 	}
 
-	def hasDefiningPoint(lon: Float, lat: Float): Boolean =
+	def hasDefiningPoint(lon: Double, lat: Double): Boolean =
 		left.sameAs(lon, lat) || top.sameAs(lon, lat) || right.sameAs(lon, lat) || bottom.sameAs(lon, lat)
 
-	def width: Float = right.lon - left.lon
-	def height: Float = top.lat - bottom.lat
+	def width: Double = right.lon - left.lon
+	def height: Double = top.lat - bottom.lat
 
 	override def toString = s"BBox($left, $top, $right, $bottom)"
 }
 
-case class Point(lon: Float, lat: Float){
+case class Point(lon: Double, lat: Double){
 
-	def sameAs(lon2: Float, lat2: Float): Boolean = lon2 == lon && lat2 == lat
+	def sameAs(lon2: Double, lat2: Double): Boolean = lon2 == lon && lat2 == lat
 
 	override def toString = s"($lon, $lat)"
 }
